requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Google ADK & Gemini
google-adk>=0.2.0
google-genai>=1.0.0

# HTTP Client (for GUVI callback)
httpx>=0.26.0

# Session Storage (optional - can use in-memory)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
python-dateutil>=2.8.0
